BulletinBoard/requirements.txt
Django>=5.0.6
django-allauth>=0.61.1
django-ckeditor>=6.7.1
celery>=5.3.6
redis>=5.0.1
django-celery-beat>=2.5.0 